requirements.txt
pandas
numpy
tensorflow
matplotlib
scikit-learn
pytorch
keras
seaborn
mlflow
flask